Ingredient Search
Added functionality to allow users to both search using a recipe name and ingredient name in the recipe recommendations search bar using the TheMealDB API.

Additionally addressed an error when fetching for links in individual recipe items. Because the URL validation is failing when the fetch request encounters network issues (like CORS, timeouts, or unreachable servers).