Added Grid Toggle
Created a toggle that allows users to choose whether or not they want their pantry list to be in a list or in a grid. In list mode, the users will have to swipe to delete items. In grid mode, they will have to hold the item in order to delete it.